Update README.md
These codes can be freely copied and pasted into OxCal 4.4 and run. Because of the random nature of the algorithm there is a chance the exact date ranges produced may alter very slightly after each run. Please cite the correct article (Kuitems et al. 2021) if and when these models are referred to in any publication.